Implement calls and refactor a bit
Instead of forcing expressions of some type, use a concept of TypeGroup,
where the generators must just generate some expression contained in the
group (which could be a single type). This allows us to better adjust
things; for example, when we want to generate a call that returns an int
type, we can generate any call that returns an integral type and cast
it.

Additionally move literal generation into its own class, and use it from
StaticsManager to generate initialization values too.