Almost finished GenBinary and fixed some bugs
Implement most of GenBinary. What is missing is left and right shifts.
These are missing because we use a table to figure out which type
combinations are valid, and these require a different table than the
other binary operators (>=, >, <, <=, ==, !=, +, -, *, /, %, ^). The
table is generated by a new project.

Fix the literal generation; we would never generate anything except 0,
MinValue, MinValue + 1, MaxValue and MaxValue - 1 in the. Fix this.
Also, fix a case where a literal would not be the correct type because
the value is assigned as an int.