Initial version of automatic reducer
This adds a --reduce switch which in combination with --seed allows to
reduce programs automatically. The reducer works by figuring out what
the problem was, and then repeatedly trying to simplify nodes by some
simple (and mostly local) rewriter rules. If the new produced tree still
has the bug, the tree is updated. This is then iterated until we reach a
fix point.