Track ref escape per scope and improve lvalue gen
Before we could generate code like:
int a = 0;
ref int ra = ref a;
{
    int c = 0;
    ra = ref c;
}

This is fixed now. We track the scope of variables by using a negative
ref escape scope, which is set up based on the count of scopes when a
variable is created.

In addition, fix ref escape scope of calls. For instance:
ref int a = ref M();
has global scope, but was being tracked as minimum value scope. Fix this
by keeping track of the arguments generated for the call, and taking the
minimum of those.

Additionally improve generation of lvalues. We now use a table and
prefer locals > statics > function calls. Also introduce a probability
for GenMemberAccess to pick locals vs statics. Previously we would end
up with a ton of statics and barely any locals, so we would almost
always pick statics. Now we prefer locals.

Also fix a bug with lvalue gen; we were wrongly returning no candidates
when we should have returned candidates of any type.