Improve simplification iters
Instead of iterating to 2000 always, check if there are fewer than 2000
nodes and iterate to the list count instead. Then instead of picking
random indices on every iteration, just start from a random index and
pick successive indices after that. This means that with fewer than 2000
nodes we just check all possible nodes and then break, which is faster
and better.