Make reducer more deterministic and improve debugging
The previous reducer optimization that changed it to only iterate a few
times when the list was shorter than the iterations was wrong. This was
due to the fact that some simplifications used randomization to select a
single candidate out of many. Change the reducer to allow simplifiers to
return an enumerator, and change its simplifier to return all possible
candidate simplifications. Then the reducer can try each of these.

Also change the simplifiction to iterate the list in a random order when
finding nodes to simplify, instead of just randomizing the start index.
This makes the simplifier much, much faster (approximately 100%
speedup).

To aid in debugging the reducer, add a --reduce-debug-git-dir switch.
When specified, Fuzzlyn will create a directory at this location and use
Git to version each update of the example being reduced. This can then
be examined to determine the order of simplifications and how the
reduction happened.

Add timing of reduction and output this as a comment as well.