Improve example checksumming
We only need the value at each checksum site when we want to comment
with the example output after reduction. Up until that point we can just
use the checksums only, so turn off the tracking. This saves memory and
CPU time when we generate very large examples.

Change the hash algorithm to 64-bit FNV-1a instead of SHA256. This is
much faster.