Improve tracking of number of calls
Instead of tracking the number of max number of leaf calls on a single
path, track it on all paths instead. The other causes problems in the
situations like:
M1() { M2(); M3(); }
M2() { M3(); M4(); }
M3() { M5(); M6(); }
where we would see a very small amount of calls, even though this also
suffers from exponential blow-up. We now track the transitive call
counts fully. Since this is not just a single path anymore, we update
the cap to get us to about where we were before in terms of examples.