Support examples that crash the runtime
Previously we only supported these examples if they happened during
reduction. Now we support them if we straight up find them during
generation. This is done by checking if the sub process crashed, and if
so, checking all examples separately. Additionally the reducer supports
examples that crash by performing the initial run in a sub process.