Rearchitecture test execution
We now use an execution server instead that stays up and running and can
be reused (unless the test example crashes). The benefit is that we
always reduce in sub processes and that we can avoid running all the
example generations in the possibly checked/debug runtime, which is very
slow. I was seeing 10s of seconds taken to serialize JSON to send to the
sub process when using the checked runtime.