Use ETW to get original exceptions
It is possible that we throw an exception and then throw another
exception in a finally being executed due to that exception. If that
happens, we would see the exception thrown in the finally and lose the
original exception -- but the original exception is the one we want to
show since it might be the cause of the subsequent one.