Fix incorrect "Output" string in reduced examples
For examples that threw we would always print the thrown exception even
if both cases threw the same exception. In reality, in most of these
cases we have a different number of lines output, which no longer was
getting tracked after the execution server (since we do not return the
checksum sites through JSON). Fix both of these issues.