Avoid mutating through reinterpreted pointers
We only reinterpret A to B if B subsets A (i.e. no field of B touches
padding in A). However, in some cases we would end up mutating through
the new pointer, which is not ok unless the opposite is also true.
Forcefully insert a ReadUnaligned to avoid this.